requirements.txt
argparse==1.2.1
oauthlib==1.0.3
python-twitter==2.2
requests==2.7.0
requests-oauthlib==0.5.0
wsgiref==0.1.2
